refactor: back to original folder sctructure
