--- src/modules/teams/index.ts
@@ -1,2 +1,0 @@
-export { TeamOverview } from './routes/TeamOverview';
-export { Teams } from './routes/Teams';
