--- src/modules/users/index.ts
@@ -1,1 +1,0 @@
-export { UserOverview } from './routes/UserOverview';
