--- src/modules/users/types/index.ts
@@ -1,8 +1,0 @@
-export interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  displayName: string;
-  location: string;
-  avatar: string;
-}
