+++ src/pages/teams/index.ts
@@ -1,0 +1,2 @@
+export { TeamOverview } from './TeamOverview';
+export { Teams } from './Teams';
