+++ src/pages/users/index.ts
@@ -1,0 +1,1 @@
+export { UserOverview } from './UserOverview';
