+++ src/types/index.ts
@@ -8,3 +8,12 @@ export interface TeamOverview {
   teamLeadId: string;
   teamMemberIds: string[];
 }
+
+export interface User {
+  id: string;
+  firstName: string;
+  lastName: string;
+  displayName: string;
+  location: string;
+  avatar: string;
+}
